requirements.txt
# ── Data ingestion ────────────────────────────────────────────────────────────
yfinance>=0.2.40
fredapi>=0.5.1
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=15.0.0
requests>=2.31.0

# ── Technical indicators ──────────────────────────────────────────────────────
ta>=0.11.0

# ── NLP / FinBERT (Phase 2) ───────────────────────────────────────────────────
transformers>=4.40.0
torch>=2.2.0
spacy>=3.7.0
newsapi-python>=0.2.7

# ── ML / modelling (Phases 3-9) ───────────────────────────────────────────────
scikit-learn>=1.4.0
lightgbm>=4.3.0
xgboost>=2.0.0
hmmlearn>=0.3.2
shap>=0.45.0

# ── Causal inference (Phase 4) ────────────────────────────────────────────────
dowhy>=0.11.1

# ── Reinforcement learning (Phases 8, 10) ─────────────────────────────────────
stable-baselines3>=2.3.0
pettingzoo>=1.24.0
gymnasium>=0.29.0

# ── Meta-learning (Phase 7) ───────────────────────────────────────────────────
higher>=0.2.1

# ── MLOps ─────────────────────────────────────────────────────────────────────
mlflow>=2.12.0
apache-airflow>=2.9.0

# ── Serving (Phase 13) ────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# ── Drift detection (Phase 13) ────────────────────────────────────────────────
evidently>=0.4.30

# ── Dashboard (Phase 12) ──────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-cov>=5.0.0

# ── Config + utilities ────────────────────────────────────────────────────────
pyyaml>=6.0.1
python-dotenv>=1.0.1
tqdm>=4.66.0
